Fixed constant re-rendering, optimized save Todo logic
